The project structure converted to package-mode.
README, setup.py and build requirements.
