+++ .github/workflows/linterandtests.yml
@@ -57,6 +57,5 @@ jobs:
       uses: codecov/codecov-action@v1
       with:
         flags: unittests
-        env_vars: OS,PYTHON
         name: codecov-umbrella
         fail_ci_if_error: true
