+++ pytest.ini
@@ -1,5 +1,5 @@
 [pytest]
-addopts = -v --flake8 --cov=armory --cov-report=xml
+addopts = -v --flake8 --cov=yapam --cov-report=xml
 testpaths = tests/
 markers =
     config
